+++ src/htsback.c
@@ -45,6 +45,7 @@ Please visit our Website: http://www.httrack.com
 
 #include "htsftp.h"
 #include "htscodec.h"
+#include "htsproxy.h"
 #if HTS_USEZLIB
 #include "htszlib.h"
 #else
